fix: функции выполняющие проверку объекта User вынесены в отдельный класс
