+++ src/main/java/ru/yandex/practicum/filmorate/servise/ValidateFilm.java
@@ -1,0 +1,4 @@
+package ru.yandex.practicum.filmorate.servise;
+
+public class ValidateFilm {
+}
